move methods into mkmatter.rb
