+++ mkmatter.gemspec
@@ -1,6 +1,6 @@
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'methods'
+require 'mkmatter'
 Gem::Specification.new do |spec|
   spec.name = Mkmatter::GEM_NAME
   spec.version = Mkmatter::VERSION
